Move FormControlElementInspector.js below lib/dom/
