+++ lib/dom/FormElementInspector.js
@@ -1,4 +1,4 @@
-import FormControlElementInspector from '../domClerks/FormControlElementInspector.js'
+import FormControlElementInspector from './FormControlElementInspector.js'
 
 import HashBuilder from '../domClerks/HashBuilder.js'
