+++ lib/domClerks/BaseLegacyFormElementClerk.js
@@ -1,4 +1,4 @@
-import FormControlElementInspector from './FormControlElementInspector.js'
+import FormControlElementInspector from '../dom/FormControlElementInspector.js'
 import VariablesValidator from '../validator/ValueHashValidator.js'
 
 /**
